Fix build process and scripts
