+++ .gitignore
@@ -12,3 +12,4 @@ tmp-tokens
 /polaris-react/build
 /polaris-react/build-internal
 /polaris-react/src/styles/_media-queries.scss
+/polaris-icons/metadata.json
