+++ polaris-icons/.babelrc
@@ -1,10 +1,6 @@
-module.exports = function (api) {
-  api.cache(true);
-
-  return {
-    presets: [
-      ['@shopify/babel-preset/web', {modules: 'auto', typescript: true}],
-      ['@shopify/babel-preset/react'],
-    ],
-  };
-};
+{
+  "presets": [
+    ["@shopify/babel-preset/web", {"modules": "auto", "typescript": true}],
+    ["@shopify/babel-preset/react"]
+  ]
+}
